requirements.txt
# Core Data Science Libraries
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
prophet>=1.1.4

# Deep Learning
tensorflow>=2.13.0
keras>=2.13.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Dashboard
streamlit>=1.25.0

# Configuration
pyyaml>=6.0

# Utilities
joblib>=1.3.0
tqdm>=4.65.0

# Data Processing
python-dateutil>=2.8.2

# Chatbot Dependencies
groq>=0.4.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
requests>=2.31.0